refactoring and moving things to be slimmer.
